% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{agenda}
\alias{agenda}
\title{Knit R Markdown to Agenda PDF}
\usage{
agenda(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Args passed to \code{rmarkdown::\link{pdf_document}()}.}

\item{keep_tex}{Is TRUE by default. Set FALSE to discard intermediate .tex file during knit}
}
\description{
Render Meeting Agenda as a PDF using the custom .tex file
}
