% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{agenda_doc}
\alias{agenda_doc}
\title{Knit R Markdown to Agenda PDF}
\usage{
agenda_doc(...)
}
\arguments{
\item{...}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}
}
\description{
Render Meeting Agenda as a PDF using the custom .tex file
}
