% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{agenda_pdf}
\alias{agenda_pdf}
\title{Knit R Markdown to Agenda PDF}
\usage{
agenda_pdf(...)
}
\arguments{
\item{...}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}

\item{md_extensions}{Markdown extensions passed to Pandoc ("+" precursor indicates include, "-" precursor indicates exclude)}
}
\description{
Render Meeting Agenda as a docx using the custom .docx file
}
