% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_template}
\alias{find_template}
\title{A Helper function used for locating a LaTeX template file.}
\usage{
find_template(template, file)
}
\arguments{
\item{template}{Name of target document template}

\item{file}{Name of the target template for this document}
}
\description{
Locates the custom .tex template file in the rmarkdown/templates directory
}
