% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutes.R
\name{minutes}
\alias{minutes}
\title{Knit R Markdown to Minutes PDF}
\usage{
minutes(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Args passed to \code{rmarkdown::\link{pdf_document}()}.}

\item{keep_tex}{Is TRUE by default. Set FALSE to discard intermediate .tex file during knit}
}
\description{
Render Meeting Minutes as a PDF using the custom .tex file
}
