% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdf_document_format}
\alias{pdf_document_format}
\title{A Helper function to create a custom word format, based on word_document.}
\usage{
pdf_document_format(template_name, ...)
}
\arguments{
\item{template_name}{Name of target document template}

\item{...}{Arguments passed to \code{rmarkdown::\link{word_document_format}()}.}
}
\description{
Locates the custom .tex template file in the rmarkdown/templates directory
}
