% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word_document_format}
\alias{word_document_format}
\title{A Helper function to create a custom word format, based on word_document.}
\usage{
word_document_format(template_name,
  docx_template = find_resource(template_name, "style-reference.docx"),
  ...)
}
\arguments{
\item{template_name}{Name of target document template}

\item{docx_template}{Name of the target docx template for this document}

\item{...}{Arguments passed to \code{rmarkdown::\link{word_document_format}()}.}

\item{pandoc_args}{Arguments passed to Pandoc}
}
\description{
Locates the custom .tex template file in the rmarkdown/templates directory
}
