% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{agenda_pdf}
\alias{agenda_pdf}
\title{Knit R Markdown to Agenda PDF}
\usage{
agenda_pdf(keep_tex = TRUE, ...)
}
\arguments{
\item{keep_tex}{Is TRUE by default. Set FALSE to discard intermediate .tex file during knit}

\item{...}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}
}
\description{
Render Meeting Agenda as a docx using the custom .docx file
}
