% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_template}
\alias{find_template}
\title{A Helper function used for locating a LaTeX template file.}
\usage{
find_template(template_file)
}
\arguments{
\item{template_file}{Name of target template file}
}
\description{
Locates the custom .tex template file in the rmarkdown/templates directory
}
